Send zync token to authenticate to System
Mutual authentication with system.
It uses the same token that System workers to authenticate with Zync.
Zync will query system with this token in the header to identify itself

Ideas could be to use BasicAuth but we would need to modify System code base.
Can be improved later I think