introduced version number in cache key
this allows us to be able to invalidate the cache (by incrementing the number), in cases like the failing build on the previous commit.

That build tries to restore a cache on the older /home/circleci/... path, whereas with the new builder image we're using in this PR, that path no longer exists.

The fix is really quite simple: don't use the cache. And this is how we can do that.